Move json utilities to their own module.
